Extracted DiagnosticParser and DiagnosticPublisher out of CargoManager.
