+++ src/components/cargo/file_diagnostic.ts
@@ -1,0 +1,7 @@
+import { Diagnostic } from 'vscode';
+
+export interface FileDiagnostic {
+    filePath: string;
+
+    diagnostic: Diagnostic;
+}
